Separated errorsPlugin and validationPlugin
