+++ packages/roqueform/src/main/index.ts
@@ -4,6 +4,7 @@
 
 export { composePlugins } from './composePlugins';
 export { createField } from './createField';
+export { ClearErrorsOptions, ErrorsPlugin, errorsPlugin } from './errorsPlugin';
 export { naturalValueAccessor } from './naturalValueAccessor';
 export {
   Field,
@@ -17,4 +18,4 @@ export {
   ValueAccessor,
 } from './types';
 export { dispatchEvents, isEqual, callOrGet } from './utils';
-export { Validator, Validation, ValidationPlugin, ValidationPluginOptions, validationPlugin } from './validationPlugin';
+export { Validator, Validation, ValidationPlugin, validationPlugin } from './validationPlugin';
